requirement.txt
# Core LangGraph packages
langgraph>=0.0.19
langchain>=0.1.0
langchain-core>=0.1.10

# LLM providers and APIs
openai>=1.6.0
anthropic>=0.8.0
cohere>=4.37
google-generativeai>=0.3.0 
langchain-google-genai>=0.0.5 
langchain-openai>=0.0.2

# Vector stores and databases
faiss-cpu>=1.7.4
chromadb>=0.4.18
langchain-community>=0.0.13
pinecone-client>=2.2.4
redis>=5.0.1
sqlalchemy>=2.0.23

# Data processing and analysis
numpy>=1.24.0
pandas>=2.1.1
scipy>=1.11.3
scikit-learn>=1.3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.29.0

# Natural Language Processing
transformers>=4.36.0
tokenizers>=0.15.0
spacy>=3.7.2
nltk>=3.8.1
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4

# Embeddings and semantic search
instructor>=0.4.5
pydantic>=2.5.2

# Document processing
unstructured>=0.12.3
pypdf>=3.17.1
docx2txt>=0.8
beautifulsoup4>=4.12.2
html2text>=2020.1.16

# Evaluation and testing
ragas>=0.1.0
deepeval>=0.20.12
pytest>=7.4.3

# Tools and agents
langchain-experimental>=0.0.42
dspy-ai>=2.3.8
crewai>=0.28.0

# Web and API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic-settings>=2.1.0
httpx>=0.25.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
jupyter>=1.0.0
ipykernel>=6.26.0